major refactor initial commit
