+++ .gitignore
@@ -1,13 +1,14 @@
-/tmp
-/dev
-/doc
-/coverage
-/gems
-/Gemfile.lock
 *.gem
 *.sublime*
-
-/snaps
-/snapcrawl.pstore
-/log.txt
+/cache
+/coverage
+/debug.rb
+/dev
+/doc
 /err.txt
+/Gemfile.lock
+/gems
+/log.txt
+/snapcrawl.pstore
+/snaps
+/tmp
