+++ Gemfile
@@ -3,12 +3,13 @@ source "https://rubygems.org"
 gemspec
 
 group :development, :test do
+  gem 'byebug'
+  gem 'lp'
+  gem 'rspec'
+  gem 'rspec_fixtures'
   gem 'runfile'
+  gem 'runfile-tasks'
+  gem 'simplecov'
   gem 'sinatra'
   gem 'sinatra-contrib'
-  gem 'runfile-tasks'
-  gem 'rspec'
-  gem 'simplecov'
-  gem 'byebug'
-  gem 'rspec_fixtures'
 end
