+++ lib/snapcrawl/exceptions.rb
@@ -1,4 +1,5 @@
 module Snapcrawl
   class MissingPhantomJS < StandardError; end
   class MissingImageMagick < StandardError; end
+  class ScreenshotError < StandardError; end
 end
