+++ lib/snapcrawl/page.rb
@@ -1,0 +1,105 @@
+require 'addressable/uri'
+require 'fileutils'
+require 'httparty'
+require 'lightly'
+require 'nokogiri'
+
+module Snapcrawl
+  class Page
+    using StringRefinements
+
+    attr_reader :url, :depth
+
+    EXTENSION_BLACKLIST = "png|gif|jpg|pdf|zip"
+    PROTOCOL_BLACKLIST = "mailto|tel"
+
+    def initialize(url, depth: 0)
+      @url, @depth = url.protocolize, depth
+    end
+
+    def http_response
+      @http_response ||= cache.get(url) do
+        HTTParty.get url
+      end
+    end
+
+    def valid?
+      http_response.success?
+    end
+
+    def warnings
+      @warnings ||= []
+    end
+
+    def pages
+      return nil unless valid?
+      links.map { |link| Page.new link, depth: depth+1 }
+    end
+
+    def links
+      return nil unless valid?
+      doc = Nokogiri::HTML http_response.body
+      normalize_links doc.css('a')
+    end
+
+    def site
+      @site ||= Addressable::URI.parse(url).site
+    end
+
+    def path
+      @path ||= Addressable::URI.parse(url).request_uri
+    end
+
+    def save_screenshot(outfile, width: 0, height: 1280, selector: nil)
+      screenshot = Screenshot.new url, height: height, width: width,
+        selector: selector
+
+      screenshot.save "#{outfile}"
+    end
+
+  private
+
+    def normalize_links(links)
+      result = []
+      @warnings = nil
+
+      links.each do |link|
+        valid_link = normalize_link link
+        result << valid_link if valid_link
+      end
+
+      result.uniq
+    end
+
+    def normalize_link(link)
+      link = link.attribute('href').to_s.dup
+
+      # Remove #hash
+      link.gsub!(/#.+$/, '')
+      return nil if link.empty?
+
+      # Remove links to specific extensions and protocols
+      return nil if link =~ /\.(#{EXTENSION_BLACKLIST})(\?.*)?$/
+      return nil if link =~ /^(#{PROTOCOL_BLACKLIST}):/
+
+      # Strip spaces
+      link.strip!
+
+      # Convert relative links to absolute
+      begin
+        link = Addressable::URI.join(url, link).to_s.dup
+      rescue => e
+        warnings << { link: link, message: "#{e.class} #{e.message}" }
+        return nil
+      end
+
+      # Keep only links in our base domain
+      return nil unless link.include? site
+      link
+    end
+
+    def cache
+      Lightly.new life: '1d'
+    end
+  end
+end
