+++ spec/server/server.rb
@@ -1,7 +1,7 @@
 require 'sinatra'
 require "sinatra/reloader" if development?
 
-set :port, 4567
+set :port, 3000
 set :bind, '0.0.0.0'
 
 before do
